+++ runtime/bin/socket_unsupported.cc
@@ -17,6 +17,12 @@ void FUNCTION_NAME(InternetAddress_Parse)(Dart_NativeArguments args) {
 }
 
 
+void FUNCTION_NAME(NetworkInterface_ListSupported)(Dart_NativeArguments args) {
+  Dart_ThrowException(DartUtils::NewDartArgumentError(
+      "Sockets unsupported on this platform"));
+}
+
+
 void FUNCTION_NAME(Socket_CreateConnect)(Dart_NativeArguments args) {
   Dart_ThrowException(DartUtils::NewDartArgumentError(
       "Sockets unsupported on this platform"));